+++ app/index.js
@@ -391,7 +391,7 @@ module.exports = generators.Base.extend({
       if (!this.bluemix) return;
       var done = this.async();
 
-      var choices = ['Cloudant', 'Redis', 'Object Storage', 'AppID', 'Auto-scaling', 'Watson Conversation', 'Alert Notification'];
+      var choices = ['Cloudant', 'Redis', 'Object Storage', 'AppID', 'Auto-scaling', 'Watson Conversation', 'Alert Notification', 'Push Notifications'];
 
       var prompts = [{
         name: 'services',
@@ -418,6 +418,9 @@ module.exports = generators.Base.extend({
         }
         if (answers.services.indexOf('Alert Notification') !== -1) {
           this._addService('alertnotification',  { name: generateServiceName(this.appname, 'AlertNotification') });
+        }
+        if (answers.services.indexOf('Push Notifications') !== -1) {
+          this._addService('pushnotifications',  { name: generateServiceName(this.appname, 'PushNotifications') });
         }
         if (answers.services.indexOf('Auto-scaling') !== -1) {
           this.autoscale = generateServiceName(this.appname, 'AutoScaling');
@@ -495,6 +498,7 @@ module.exports = generators.Base.extend({
           case 'appid':               return 'AppID';
           case 'watsonconversation':  return 'Watson Conversation';
           case 'alertnotification':  return 'Alert Notification';
+          case 'pushnotifications':  return 'Push Notifications';
           default:
             self.env.error(chalk.red(`Internal error: unknown service type ${serviceType}`));
         }
@@ -649,6 +653,46 @@ module.exports = generators.Base.extend({
           password: answers.alertNotificationPassword || undefined,
           url: answers.alertNotificationUrl || undefined
         };
+        done();
+      }.bind(this));
+    },
+
+    promptConfigurePushNotifications: function() {
+      if (this.skipPrompting) return;
+      if (!this.servicesToConfigure) return;
+      if (!this.servicesToConfigure.pushnotifications) return;
+      var done = this.async();
+
+      this.log();
+      this.log('Configure Push Notifications');
+      
+      var prompts = [
+        { name: 'pushNotificationsName', message: 'Enter service name (blank for default):',
+          when: (answers) => this.bluemix
+        },
+        { name: 'pushNotificationsAppGuid', message: 'Enter app GUID:' },
+        { name: 'pushNotificationsAppSecret', message: 'Enter app secret:', type: 'password' },
+        {
+          name: 'pushNotificationsRegion',
+          type: 'list',
+          message: 'Enter Bluemix region:',
+          choices: [ 'US South', 'United Kingdom', 'Sydney' ],
+          default: 'US South'
+        }
+      ];
+      this.prompt(prompts, function(answers) {
+        this.services.pushnotifications[0].name = answers.pushNotificationsName || this.services.pushnotifications[0].name;
+        this.services.pushnotifications[0].credentials = {
+          appGuid: answers.pushNotificationsAppGuid || undefined,
+          appSecret: answers.pushNotificationsAppSecret || undefined
+        };
+        switch (answers.pushNotificationsRegion) {
+          case 'US South':        this.services.pushnotifications[0].region = 'US_SOUTH'; break;
+          case 'United Kingdom':  this.services.pushnotifications[0].region = 'UK'; break;
+          case 'Sydney':          this.services.pushnotifications[0].region = 'SYDNEY'; break;
+          default:
+            this.env.error(chalk.red(`Internal error: unknown region ${answers.pushNotificationsRegion}`));
+        }
         done();
       }.bind(this));
     },