+++ refresh/templates/services/pushnotifications/importDependency.swift
@@ -1,0 +1,1 @@
+        .Package(url: "https://github.com/ibm-bluemix-mobile-services/bms-pushnotifications-serversdk-swift.git", majorVersion: 0, minor: 5),